refactor : move review domain
