+++ backend/src/main/java/toonpick/app/dto/WebtoonReviewCreateDTO.java
@@ -1,4 +1,4 @@
-package toonpick.app.review.dto;
+package toonpick.app.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
