+++ backend/src/main/java/toonpick/app/dto/WebtoonReviewDTO.java
@@ -1,10 +1,9 @@
-package toonpick.app.review.dto;
+package toonpick.app.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import toonpick.app.dto.MemberDTO;
 
 import java.time.LocalDateTime;
