+++ backend/src/main/java/toonpick/app/repository/WebtoonReviewRepository.java
@@ -1,4 +1,4 @@
-package toonpick.app.review.repository;
+package toonpick.app.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 import toonpick.app.domain.member.Member;
 import toonpick.app.domain.webtoon.Webtoon;
-import toonpick.app.review.entity.WebtoonReview;
+import toonpick.app.domain.review.WebtoonReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
