+++ backend/src/test/java/toonpick/app/unit/service/WebtoonReviewServiceTest.java
@@ -5,18 +5,18 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import toonpick.app.review.dto.WebtoonReviewCreateDTO;
-import toonpick.app.review.dto.WebtoonReviewDTO;
+import toonpick.app.dto.WebtoonReviewCreateDTO;
+import toonpick.app.dto.WebtoonReviewDTO;
 import toonpick.app.domain.member.Member;
 import toonpick.app.domain.webtoon.Webtoon;
-import toonpick.app.review.entity.WebtoonReview;
+import toonpick.app.domain.review.WebtoonReview;
 import toonpick.app.domain.webtoon.enums.Platform;
 import toonpick.app.domain.webtoon.enums.SerializationStatus;
 import toonpick.app.exception.ResourceNotFoundException;
 import toonpick.app.repository.MemberRepository;
 import toonpick.app.repository.WebtoonRepository;
-import toonpick.app.review.repository.WebtoonReviewRepository;
-import toonpick.app.review.service.WebtoonReviewService;
+import toonpick.app.repository.WebtoonReviewRepository;
+import toonpick.app.service.WebtoonReviewService;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
